Clean up bicep and refactor app
