+++ samples/portable-sdks/dotnet/FanOutFanIn/OrchestrationService/Models/OrchestrationResultModels.cs
@@ -1,0 +1,9 @@
+namespace OrchestrationService.Models;
+
+public class FanOutFanInTestResult
+{
+    public int TotalActivities { get; set; }
+    public long ElapsedTimeMs { get; set; }
+    public double AverageActivityTimeMs { get; set; }
+    public List<ActivityResult> Results { get; set; } = new List<ActivityResult>();
+}
