+++ samples/portable-sdks/dotnet/FanOutFanIn/infra/main.bicep
@@ -12,17 +12,40 @@ param location string
 @description('Id of the user or app to assign application roles')
 param principalId string = ''
 
+// ACR parameters
+@description('Name of the Azure Container Registry (ACR)')
+param acrName string = ''
+
 var tags = {
   'azd-env-name': environmentName
 }
 
+// Use unique naming for ACR
+var uniqueAcrName = !empty(acrName) ? acrName : 'acr${uniqueString(subscription().id, environmentName)}'
+
 // Role definition ID for "Durable Task Data Contributor"
 var durableTaskDataContributorRoleId = '0ad04412-c4d5-4796-b79c-f76d14c8d402'
+
+// Role definition ID for "AcrPush" - allows pushing images to ACR
+var acrPushRoleId = '8311e382-0749-4cb8-b61a-304f252e45ec'
+// Role definition ID for "AcrPull" - allows pulling images from ACR
+var acrPullRoleId = '7f951dda-4ed3-4680-a7ca-43fe172d538d'
 
 resource resourceGroup 'Microsoft.Resources/resourceGroups@2021-04-01' = {
   name: '${environmentName}-rg'
   location: location
   tags: tags
+}
+
+// Azure Container Registry for storing application images
+module containerRegistry 'modules/container-registry.bicep' = {
+  name: 'container-registry'
+  scope: resourceGroup
+  params: {
+    name: uniqueAcrName
+    location: location
+    tags: tags
+  }
 }
 
 // Container Apps Environment and required services
@@ -53,7 +76,7 @@ module durableTaskScheduler 'modules/durable-task-scheduler.bicep' = {
 
 // User assigned managed identities for services
 module orchestrationServiceIdentity 'modules/managed-identity.bicep' = {
-  name: 'orchestration-service-identity'
+  name: 'orchestration-service-identity-module'
   scope: resourceGroup
   params: {
     name: '${environmentName}-orchestration-identity'
@@ -63,7 +86,7 @@ module orchestrationServiceIdentity 'modules/managed-identity.bicep' = {
 }
 
 module workerServiceIdentity 'modules/managed-identity.bicep' = {
-  name: 'worker-service-identity'
+  name: 'worker-service-identity-module'
   scope: resourceGroup
   params: {
     name: '${environmentName}-worker-identity'
@@ -109,7 +132,43 @@ module dashboardDtsAccess 'modules/dts-access.bicep' = if (!empty(principalId)) 
   }
 }
 
-// Application services
+// ACR Push access for the current user (if principalId is provided)
+module userAcrPushAccess 'modules/role-assignment.bicep' = if (!empty(principalId)) {
+  name: 'user-acr-push-access'
+  scope: resourceGroup
+  params: {
+    principalId: principalId
+    roleDefinitionId: acrPushRoleId
+    principalType: 'User'
+    resourceId: containerRegistry.outputs.id
+  }
+}
+
+// ACR Pull access for orchestration service
+module orchestrationServiceAcrPullAccess 'modules/role-assignment.bicep' = {
+  name: 'orchestration-service-acr-pull'
+  scope: resourceGroup
+  params: {
+    principalId: orchestrationServiceIdentity.outputs.principalId
+    roleDefinitionId: acrPullRoleId
+    principalType: 'ServicePrincipal'
+    resourceId: containerRegistry.outputs.id
+  }
+}
+
+// ACR Pull access for worker service
+module workerServiceAcrPullAccess 'modules/role-assignment.bicep' = {
+  name: 'worker-service-acr-pull'
+  scope: resourceGroup
+  params: {
+    principalId: workerServiceIdentity.outputs.principalId
+    roleDefinitionId: acrPullRoleId
+    principalType: 'ServicePrincipal'
+    resourceId: containerRegistry.outputs.id
+  }
+}
+
+// Application services - Orchestration Service
 module orchestrationService 'modules/container-app.bicep' = {
   name: 'orchestration-service'
   scope: resourceGroup
@@ -118,11 +177,9 @@ module orchestrationService 'modules/container-app.bicep' = {
     location: location
     tags: tags
     environmentId: containerAppsEnvironment.outputs.id
-    containerImage: 'orchestration-service:latest'
+    containerImage: '${containerRegistry.outputs.loginServer}/orchestration-service:latest'
     containerPort: 8080
-    containerRegistry: ''
-    containerRegistryUsername: ''
-    containerRegistryPassword: ''
+    containerRegistry: containerRegistry.outputs.loginServer
     userAssignedIdentityId: orchestrationServiceIdentity.outputs.id
     env: [
       {
@@ -138,6 +195,7 @@ module orchestrationService 'modules/container-app.bicep' = {
   }
 }
 
+// Application services - Worker Service
 module workerService 'modules/container-app.bicep' = {
   name: 'worker-service'
   scope: resourceGroup
@@ -146,11 +204,9 @@ module workerService 'modules/container-app.bicep' = {
     location: location
     tags: tags
     environmentId: containerAppsEnvironment.outputs.id
-    containerImage: 'worker-service:latest'
+    containerImage: '${containerRegistry.outputs.loginServer}/worker-service:latest'
     containerPort: 8080
-    containerRegistry: ''
-    containerRegistryUsername: ''
-    containerRegistryPassword: ''
+    containerRegistry: containerRegistry.outputs.loginServer
     userAssignedIdentityId: workerServiceIdentity.outputs.id
     env: [
       {
@@ -168,4 +224,6 @@ module workerService 'modules/container-app.bicep' = {
 
 output ORCHESTRATION_SERVICE_URI string = orchestrationService.outputs.uri
 output WORKER_SERVICE_URI string = workerService.outputs.uri
+output AZURE_CONTAINER_REGISTRY_ENDPOINT string = containerRegistry.outputs.loginServer
+output AZURE_CONTAINER_REGISTRY_NAME string = containerRegistry.outputs.name
 output AZURE_LOCATION string = location
