+++ samples/portable-sdks/dotnet/FanOutFanIn/infra/modules/container-registry.bicep
@@ -1,0 +1,29 @@
+param name string
+param location string = resourceGroup().location
+param tags object = {}
+
+@allowed([
+  'Basic'
+  'Standard'
+  'Premium'
+])
+param skuName string = 'Basic'
+
+@description('Enable admin user for the container registry')
+param adminUserEnabled bool = false
+
+resource containerRegistry 'Microsoft.ContainerRegistry/registries@2023-01-01-preview' = {
+  name: name
+  location: location
+  tags: tags
+  sku: {
+    name: skuName
+  }
+  properties: {
+    adminUserEnabled: adminUserEnabled
+  }
+}
+
+output id string = containerRegistry.id
+output loginServer string = containerRegistry.properties.loginServer
+output name string = containerRegistry.name
